requirements.txt
Django
faiss-cpu
sentence-transformers
numpy
requests
gunicorn
python-dotenv
whitenoise  